+++ app/src/main/java/com/heyzeusv/yourlists/database/models/Category.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.Ignore
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import com.heyzeusv.yourlists.database.DatabaseEntity
+import com.heyzeusv.yourlists.util.portation.CsvInfo
 
 @Entity(
     indices = [Index(
@@ -17,7 +17,7 @@ data class Category(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0L,
     val name: String = "",
-) : DatabaseEntity {
+) : CsvInfo {
     @Ignore
     override val csvName: String = this::class.java.simpleName
     @Ignore
